ZEN-4456 A few pom file adjustments. Works now.
The reason the test was failing is the same reason ic couldn't get this
to work with a the RAPIDML doc gentemplate migrated to genflow: we had
two differentv versions of the org.eclipse.equinox.common library in the
class path. I din't even know mave would do that! Two different xtext
libraries - both at v2.9.2 - were the source of the conflicting
dependencies.

I was able to fix this by suppressing v3.6.0, in favor of v3.10.0. This
fixed the test as well as the genflow execution.

We still can't use the included manifests in the jar files. The
maven-jar-plugin configuration does actually do what it advertises, but
that ends up breaking the build of com.reprezen.rapidml, for reasons I
investigated but to no avail.FWIW, the state of the working tree after
running the parent+realization+model projects (via maven's -pl commnad
line arg) was identical with and without the jar plugin
configuration (not counting stuff like timestamps). Throw in the next
build, and it breaks. For now, things to be OK with the bare-bones
generated manifests, so we can go with that.

Aos moved the `libraries` directory in the crr project to the normal
place for resources in a non-tycho project. Not finding those was the
remaining issue I had trying to use this in genflow.