Rollback xtext configuration to use old code generator
This commit re-introduces the mwe2 configuration that uses the old Xtext generator in hope to fix the formatting and serialization issues we are seeing when using the
new Xtext code generator.
This has been done by migrating the c.m.r.restapi.xtext project present in API Studio to a maven layout and by renaiming packages to follow the convention we introduced in c.r.rapidml.model and c.r.rapidml.realization. Updates have also been done to make sure the grammar is compatible with the model code that now doesn't use sub-packages.
The Xtext code generator now cannot be invoked from maven but can be invoke from within Eclipse. The paths and names in the mwe2 configuration file have been updated to make sure the code generation still work.