[ZEN-4575] Several false negatives in content assist tests
Failing tests in c.m.r.xtext.ui can be resolved by making sure unnecessary language registration are not performed during resolution of library models. This is done during scope resolution. Previously a call to the standalone setup was done each time a library model was requested, thus doing a new language registration and injector creation. This is avoided by injecting a ResourceSet inside the ScopeProvider and the DslRestModelLoader class. Injectors should be created once, e.g. when the UI plugin is activated or once when a runtime action is performed outside UI.