[ZEN-4725] GenFlow should follow Maven lifecycle, should not require install before test
This change is necesssary to allow running tests (GenFlow tests) without requiring a pre-install. The problem was in the ImportResolver that uses java.net.URI that does not support resolving archives URI (starting with jar:) and thus failed to resolve imports that are located in a jar file during run of tests in maven install phase. The necessary change was to modify the implementation of ImportResolver and ModelPathItem to use EMF URI implementation, instead of java.net.URI, which supports resolving archive URIs.