[ZEN-4725] GenFlow should follow Maven lifecycle, should not require install before test
This brings back the exception handling as it was done before the change that removed use of java.net.URI. The difference is now the exception we catch is
an IllegalArgumentException that is the exception thrown by EMF URI.createURI method. This exception can be thrown when the constructor ImporterResolver is called or
when it's method resolve is called. The exception is caught in RepreZenImportUriGlobalScopeProvider which is the only place where those methods are called. The exception is for now caught but made silent, as it was before.

This also simplifies the logic in the method getUrisForItem where the logic to handle potential NPE when containerUri is null is done in a separate method.