Upgrade runtime version of Xtext to 2.15
The mwe2 configuration requires Xtext 2.9 to generate code but the runtime can be updated to Xtext 2.15.
There was incompatible changes in the tests due to the change of the dependecy xtext.junit to xtext.testing that removed some tests helper class.
Most of the tests have been updated to xtext.testing API but for some it was easier to copy the missing helper class from previous Xtext version and run
our tests with it.

The  code generation cannot be done with Xtext 2.15. Thus this project should be developed in an Eclipse having the EMF 2.13 and Xtext 2.9 plugins. The code generator
can be then run from Eclipse itself.