Update Xtext runtime dependency to 2.12 while keeping code generator to Xtext 2.9
This is a test to see if it is possible to update the runtime dependency of RAPID-ML to Xtext 2.12 which is the Xtext version matching Eclipse Oxygen release.
We can only update the runtime version of Xtext, meaning the Xtext version that is set in the pom file dependencies. It is not possible yet to use this version to
execute the code generator for the grammar as it uses an outdated generator. Thus it is required to use 2.9 for this.

This is achieved by adding a separate sub project named c.r.rapidml.generator. This project is an Eclipse project that contains the grammars and the mwe2 configuration. This project has a dependecy on Xtext 2.9. This project is not a maven project and thus is not taken into account when a RAPID-ML build is performed. This project is solely here to allow execution of the code generator from within Eclipse by invoking the execution of the mwe2 file.